Displaying 15 cards
Some changes based on the flutter update. I have also worked on displaying 15 cards on the screen. However, I could not figure out how to randomize it. And I forgot which card series we are going to use so I just pick 15 random cards to display. Need to change the it after we have a data base for all the cards.